refactoring code to build tests
