+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/SocketClientMessageHandler.java
@@ -1,5 +1,8 @@
 package com.mageddo.dnsproxyserver.server.dns;
 
+/**
+ * Lead at treat client messages.
+ */
 public interface SocketClientMessageHandler {
 
   default void handle(SocketClient client) {
