+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/UDPServer.java
@@ -16,9 +16,9 @@ public class UDPServer {
 
   private final ExecutorService pool;
   private final SocketAddress address;
-  private final RequestHandler requestHandler;
+  private final RequestHandlerDefault requestHandler;
 
-  public UDPServer(SocketAddress address, RequestHandler requestHandler) {
+  public UDPServer(SocketAddress address, RequestHandlerDefault requestHandler) {
     this.address = address;
     this.requestHandler = requestHandler;
     this.pool = ThreadPool.create();
