+++ src/test/java/com/mageddo/dnsproxyserver/server/dns/TCPServerTest.java
@@ -1,0 +1,9 @@
+package com.mageddo.dnsproxyserver.server.dns;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TCPServerTest {
+
+}
